ml/requirements.txt
# FastAPI ML Service Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
scikit-learn==1.6.0
pandas==2.2.3
pyyaml==6.0.2
scipy==1.14.1
numpy==2.2.1
pydantic==2.10.4
python-multipart==0.0.18
